PODS-1756 : package renaming
