+++ test/models/register/SchemeTypeSpec.scala
